+++ src/main/java/ea/badge/domain/Timeslot.java
@@ -18,7 +18,7 @@ import java.time.LocalTime;
 public class Timeslot {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     @Column(name = "start_time")
     private LocalTime startTime;
     @Column(name = "end_time")
@@ -32,7 +32,7 @@ public class Timeslot {
     @Enumerated(EnumType.STRING)
     private Days days;
 
-    public Timeslot(Integer id, LocalTime startTime, LocalTime endTime) {
+    public Timeslot(Long id, LocalTime startTime, LocalTime endTime) {
         this.id = id;
         this.startTime = startTime;
         this.endTime = endTime;
