+++ src/main/java/ea/badge/service/PlanServiceImpl.java
@@ -24,7 +24,7 @@ public class PlanServiceImpl implements PlanService {
     }
 
     @Override
-    public Plan findById(Integer id) {
+    public Plan findById(Long id) {
         return planRepository.findById(id).get();
     }
 
@@ -46,7 +46,7 @@ public class PlanServiceImpl implements PlanService {
     }
 
     @Override
-    public void removePlan(Integer id) {
+    public void removePlan(Long id) {
         planRepository.deleteById(id);
     }
 }
